feat: allow specifying an initial value for a prompt
This change allows the user to specify an initial value for a prompt,
similar to how "select" prompts allow specifying the initial position of
the cursor.
If an initial value is given, it is automatically placed in the prompt,
and the cursor is moved to the end, as if the user just typed this input
himself.