Better "select" & "select_multiple" Scrolling.
Updated:
- Allows the user to press down when at the bottom of a list to send the cursor/index back to the top.
- Ditto ^^ but for when the user is at the top of a list and presses up. 

<br />

Left alone:
- I didn't touch the "confirm" function as it is just 2 options and is basically a True or False check and I like it how it is.
- I didn't touch the "prompt" function as it is unrelated.

<br />
<br />

Why?:
When you start to have pretty BIG lists, say 20 options for example...it will take some time to get all the way back to the top of the list of options. Same for when you want to get to the bottom of the list. I would have to tap the down arrow key like 19 times to get to the bottom. I can now just press up 1 time and BAM, bottom of the list. (Same for being at the bottom and getting back to the top)
__ __

<br />
<br />


# Tested with:
> Successfully. 
- Python Version: v3.10
- OS: Manjaro Linux x86_64
- Kernel: 5.15.85-1-MANJARO
__ __