Update _internals.py
Update the regex pattern and the replacement when getting a list of style strings to allow square brackets in string_w_styles.

Ignores the exception where the text in-between square brackets is not a valid style string. In this case the text is left as is, since it's not a style that needs to get updated.